feat: togglable dropdown on mobile
