+++ lib/default-theme/Sidebar.vue
@@ -87,24 +87,6 @@ function resolveOpenGroupIndex (route, items) {
       font-weight 600
       font-size 1.1em
       padding 0.5rem 0 0.5rem 1.5rem
-    .nav-item .dropdown-wrapper
-      .dropdown-title
-        display inline-block
-        margin-bottom 0.5rem
-      .nav-dropdown
-        .dropdown-item
-          h4
-            border-top 0
-            margin-top 0
-            padding-top 0
-          h4, & > a
-            font-size 15px
-            height 2rem
-            line-height 2rem
-          .dropdown-subitem
-            font-size 14px
-            padding-left 1rem
-
   .sidebar-links
     margin-top 1.5rem
